backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.21
anthropic>=0.40.0
python-dotenv==1.0.1
